29 Jan 18 updated(Rename Classes/Upload gallary to firebase)
