+++ app/src/main/java/com/example/niden/cellwatchsharing/activities/TaskDetailActivity.java
@@ -1,19 +1,14 @@
 package com.example.niden.cellwatchsharing.activities;
 
-import android.app.Activity;
-import android.content.ClipData;
+
 import android.content.Intent;
 import android.database.Cursor;
 import android.net.Uri;
-import android.os.Environment;
-import android.provider.MediaStore;
 import android.provider.OpenableColumns;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.GridLayoutManager;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.util.Log;
 import android.view.View;
 import android.widget.EditText;
 import android.widget.ImageView;
@@ -21,9 +16,9 @@ import android.widget.Toast;
 
 import com.example.niden.cellwatchsharing.R;
 import com.example.niden.cellwatchsharing.adapters.UploadListAdapter;
+import com.example.niden.cellwatchsharing.classes.Account;
 import com.example.niden.cellwatchsharing.database.GallaryEntityDatabase;
 import com.example.niden.cellwatchsharing.database.TaskEntityDatabase;
-import com.example.niden.cellwatchsharing.classes.User;
 import com.example.niden.cellwatchsharing.utils.GallaryUtils;
 import com.example.niden.cellwatchsharing.utils.KeyboardUtils;
 import com.google.android.gms.tasks.OnSuccessListener;
@@ -37,11 +32,9 @@ import com.google.firebase.storage.FirebaseStorage;
 import com.google.firebase.storage.StorageReference;
 import com.google.firebase.storage.UploadTask;
 
-import java.io.File;
-import java.io.IOException;
-import java.text.SimpleDateFormat;
+
 import java.util.ArrayList;
-import java.util.Date;
+
 import java.util.List;
 
 public class TaskDetailActivity extends AppCompatActivity {
@@ -54,7 +47,7 @@ public class TaskDetailActivity extends AppCompatActivity {
     EditText etTaskName, etClass, etDescription, etAddress, etSuburb;
     String strTaskName, strDescription, strAddress, strClass, strSuburb;
     DatabaseReference mMessagesDatabaseReference;
-    User user = new User();
+    Account account = new Account();
     TaskEntityDatabase taskEntityDatabase = new TaskEntityDatabase();
     GallaryEntityDatabase gallaryEntityDatabase;
 
@@ -89,7 +82,7 @@ public class TaskDetailActivity extends AppCompatActivity {
 
        fileNameList = new ArrayList<>();
        fileDoneList = new ArrayList<>();
-        uploadListAdapter = new UploadListAdapter(fileNameList, fileDoneList);
+       uploadListAdapter = new UploadListAdapter(fileNameList, fileDoneList);
 
         //RecyclerView
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
@@ -112,7 +105,7 @@ public class TaskDetailActivity extends AppCompatActivity {
 
     public void displayTaskDetail(final EditText etTaskName, final EditText etClass, final EditText etDescription, final EditText etAddress, final EditText etSuburb) {
         mMessagesDatabaseReference = FirebaseDatabase.getInstance().getReference("users")
-                .child(user.getFirebaseAuth().getUid())
+                .child(account.getFirebaseAuth().getUid())
                 .child("tasks")
         ;
 
@@ -120,14 +113,12 @@ public class TaskDetailActivity extends AppCompatActivity {
             @Override
             public void onDataChange(DataSnapshot dataSnapshot) {
 
-
-                    taskEntityDatabase = dataSnapshot.getValue(TaskEntityDatabase.class);
-                    strTaskName = taskEntityDatabase.getTask_name();
-                    strDescription = taskEntityDatabase.getTask_description();
-                    strAddress = taskEntityDatabase.getTask_address();
-                    strClass = taskEntityDatabase.getTask_class();
-                    strSuburb = taskEntityDatabase.getTask_suburb();
-
+                taskEntityDatabase = dataSnapshot.getValue(TaskEntityDatabase.class);
+                strTaskName = taskEntityDatabase.getTask_name();
+                strDescription = taskEntityDatabase.getTask_description();
+                strAddress = taskEntityDatabase.getTask_address();
+                strClass = taskEntityDatabase.getTask_class();
+                strSuburb = taskEntityDatabase.getTask_suburb();
 
                 etTaskName.setText(strTaskName);
                 etClass.setText(strClass);
@@ -153,7 +144,7 @@ public class TaskDetailActivity extends AppCompatActivity {
 
 
 
-
+// ACTIVITY RESULT
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         if(requestCode == RESULT_LOAD_IMAGE && resultCode == RESULT_OK){
@@ -172,7 +163,7 @@ public class TaskDetailActivity extends AppCompatActivity {
                     fileDoneList.add("uploading");
                     uploadListAdapter.notifyDataSetChanged();
 
-                    StorageReference fileToUpload = mStorage.child("Images").child(fileName);
+                    StorageReference fileToUpload = mStorage.child("Gallary").child(fileName);
 
                     final int finalI = i;
                     fileToUpload.putFile(fileUri).addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
@@ -181,7 +172,6 @@ public class TaskDetailActivity extends AppCompatActivity {
 
                             fileDoneList.remove(finalI);
                             fileDoneList.add(finalI, "done");
-
                             uploadListAdapter.notifyDataSetChanged();
 
                         }
@@ -200,6 +190,8 @@ public class TaskDetailActivity extends AppCompatActivity {
         }
     }
 
+
+    //GET FILE NAME
     public String getFileName(Uri uri) {
         String result = null;
         if (uri.getScheme().equals("content")) {
