+++ app/src/main/java/com/example/niden/cellwatchsharing/activities/TechnicianActivity.java
@@ -5,7 +5,6 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.view.animation.OvershootInterpolator;
 import android.widget.ImageView;
@@ -14,13 +13,12 @@ import android.widget.TextView;
 
 import com.example.niden.cellwatchsharing.R;
 import com.example.niden.cellwatchsharing.adapters.ListTaskAdapter;
-import com.example.niden.cellwatchsharing.classes.FireBaseRetrieve;
 import com.example.niden.cellwatchsharing.classes.User;
+import com.example.niden.cellwatchsharing.classes.Account;
 import com.example.niden.cellwatchsharing.utils.IntentUtils;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.database.FirebaseDatabase;
 import com.google.firebase.database.Query;
-import com.squareup.picasso.Picasso;
 
 import jp.wasabeef.recyclerview.animators.SlideInDownAnimator;
 import me.everything.android.ui.overscroll.OverScrollDecoratorHelper;
@@ -29,24 +27,26 @@ import static com.example.niden.cellwatchsharing.activities.MainActivity.activit
 
 public class TechnicianActivity extends AppCompatActivity {
     TextView textViewName,textViewBio,textViewPhone,textViewHobby,textViewDateBirth;
-    User user = new User();
+    Account account = new Account();
     ImageView profileImage;
     RecyclerView recyclerView;
     Query mRef;
-    FireBaseRetrieve mFirebaseRetrive = new FireBaseRetrieve();
+    User mFirebaseRetrive = new User();
     RelativeLayout cover;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_technician1);
+        setContentView(R.layout.activity_technician);
        // Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         //setSupportActionBar(toolbar);
 
-        profileImage = (ImageView)findViewById(R.id.user_profile_pic);
+
+
+        profileImage= findViewById(R.id.profile_image);
         textViewName = (TextView)findViewById(R.id.user_profile_name);
         textViewBio = (TextView)findViewById(R.id.user_profile_short_bio);
-        //textViewContact = (TextView)findViewById(R.id.user);
         recyclerView = (RecyclerView) findViewById(R.id.listTask);
         cover = (RelativeLayout)findViewById(R.id.background);
 
@@ -70,7 +70,6 @@ public class TechnicianActivity extends AppCompatActivity {
         recyclerView.getItemAnimator().setAddDuration(1000);
         recyclerView.getItemAnimator().setMoveDuration(1000);
         recyclerView.setAdapter(mAdapter);
-
         OverScrollDecoratorHelper.setUpStaticOverScroll(cover, OverScrollDecoratorHelper.ORIENTATION_VERTICAL);
 
     }
@@ -88,7 +87,7 @@ public class TechnicianActivity extends AppCompatActivity {
 
     @Override
     protected void onRestart() {
-        user.checkUserLogin(activity);
+        account.checkUserLogin(activity);
         super.onRestart();
     }
