--- app/src/main/java/com/example/niden/cellwatchsharing/classes/ProfileUser.java
@@ -1,70 +1,0 @@
-package com.example.niden.cellwatchsharing.classes;
-
-import android.app.ProgressDialog;
-import android.content.Context;
-import android.net.Uri;
-import android.support.annotation.NonNull;
-import android.widget.Toast;
-
-import com.example.niden.cellwatchsharing.database.ProfilePhotoEntityDatabase;
-import com.google.android.gms.tasks.OnFailureListener;
-import com.google.android.gms.tasks.OnSuccessListener;
-import com.google.firebase.auth.FirebaseAuth;
-import com.google.firebase.database.DatabaseReference;
-import com.google.firebase.storage.OnProgressListener;
-import com.google.firebase.storage.StorageReference;
-import com.google.firebase.storage.UploadTask;
-
-import java.util.HashMap;
-import java.util.Map;
-import java.util.UUID;
-
-/**
- * Created by niden on 26-Jan-18.
- */
-
-public class ProfileUser {
-
-    public void uploadProfilePicture(final Context context, Uri filePath, StorageReference storageReference, final DatabaseReference databaseReference){
-        if(filePath != null)
-        {
-            final ProgressDialog progressDialog = new ProgressDialog(context);
-            progressDialog.setTitle("Uploading...");
-            progressDialog.show();
-
-            StorageReference mRefStorage = storageReference.child("images/"+ UUID.randomUUID().toString());
-            mRefStorage.putFile(filePath)
-                    .addOnSuccessListener(new OnSuccessListener<UploadTask.TaskSnapshot>() {
-                        @Override
-                        public void onSuccess(UploadTask.TaskSnapshot taskSnapshot) {
-                            progressDialog.dismiss();
-                            Toast.makeText(context, "Uploaded", Toast.LENGTH_SHORT).show();
-                            Uri downloadUrl = taskSnapshot.getDownloadUrl();
-                            String url = downloadUrl.toString();
-                            Map<String, Object> user = new HashMap<>();
-                            user.put("profile_url",url);
-                            databaseReference.child("users")
-                                    .child(FirebaseAuth.getInstance().getCurrentUser().getUid())
-                                    .child("info").updateChildren(user);
-
-                        }
-                    })
-                    .addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            progressDialog.dismiss();
-                            Toast.makeText(context, "Failed "+e.getMessage(), Toast.LENGTH_SHORT).show();
-                        }
-                    })
-                    .addOnProgressListener(new OnProgressListener<UploadTask.TaskSnapshot>() {
-                        @Override
-                        public void onProgress(UploadTask.TaskSnapshot taskSnapshot) {
-                            double progress = (100.0*taskSnapshot.getBytesTransferred()/taskSnapshot
-                                    .getTotalByteCount());
-                            progressDialog.setMessage("Uploaded "+(int)progress+"%");
-                        }
-                    });
-        }
-    }
-
-}
