+++ app/src/main/res/layout/activity_technician.xml
@@ -1,100 +1,26 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:attribute="http://schemas.android.com/apk/res-auto"
+<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:orientation="vertical"
+    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@color/cardview_light_background">
+    tools:context="com.example.niden.cellwatchsharing.activities.TechnicianActivity">
 
-    <ImageView
-        android:id="@+id/header_cover_image"
+    <include
+        android:id="@+id/profile_cover"
+        layout="@layout/element_profile_cover"
         android:layout_width="match_parent"
-        android:layout_height="200dp"
-        android:background="@drawable/style_background_triangle"
-        android:layout_above="@id/header_cover_image"
-         />
+        android:layout_height="wrap_content" />
 
-    <ImageButton
-        android:id="@+id/user_profile_photo"
-        android:layout_width="120dp"
-        android:layout_height="120dp"
-        android:layout_below="@+id/header_cover_image"
-        android:layout_centerHorizontal="true"
-        android:layout_marginTop="-60dp"
-        android:background="@drawable/style_circle_profile"
-        android:elevation="1dp"
-        android:padding="20dp"
-        android:scaleType="centerCrop"
-        android:src="@drawable/ic_user_blue" />
-
-    <TextView
-        android:id="@+id/user_profile_short_bio"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/user_profile_name"
-        android:layout_centerHorizontal="true"
-        android:layout_marginBottom="25dp"
-        android:layout_marginTop="12dp"
-        android:textColor="#fff"
-        android:textSize="14sp"
-        android:textStyle="bold" />
-
-    <TextView
-        android:id="@+id/user_profile_name"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_below="@+id/btn_edit_profile"
-        android:layout_centerHorizontal="true"
-        android:textColor="#fff"
-        android:textSize="24sp"
-        android:textStyle="bold" />
-
-    <ImageView
-        android:id="@+id/btn_edit_profile"
-        android:layout_width="30dp"
-        android:layout_height="30dp"
-        android:layout_alignParentRight="true"
-        android:layout_marginLeft="16dp"
-        android:layout_marginTop="16dp"
-        android:src="@android:drawable/ic_menu_edit" />
-
-    <RelativeLayout
-        android:id="@+id/profile_layout"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_alignParentStart="true"
-        android:layout_below="@+id/header_cover_image"
-        android:background="@color/cardview_light_background"
-        android:paddingBottom="24dp">
-
-    </RelativeLayout>
-
-
-    <LinearLayout
+    <android.support.v7.widget.RecyclerView
+        android:id="@+id/listTask"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        android:layout_below="@+id/profile_layout"
-        android:orientation="horizontal"></LinearLayout>
-
-    <LinearLayout
-        android:id="@+id/linearLayout"
-        android:layout_width="fill_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@+id/profile_layout"
-        android:layout_marginLeft="25dp"
-        android:layout_marginRight="25dp"
-        android:layout_marginTop="50dp"
+        android:layout_marginLeft="20dp"
+        android:layout_marginRight="20dp"
+        android:layout_marginTop="10dp"
         android:alpha="1"
-        android:background="@drawable/ui_messagebox"
-        android:elevation="4dp"
-        android:orientation="vertical">
+        android:elevation="4dp"/>
 
-        <android.support.v7.widget.RecyclerView
-            android:id="@+id/listTask"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent"
-            android:background="@drawable/style_button_border_only" />
-    </LinearLayout>
 
-</RelativeLayout>
+</LinearLayout>
