+++ app/src/main/res/menu/create_task_fragment_menu.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:tools="http://schemas.android.com/tools"
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:icon="@drawable/ic_menu_send"
+        android:iconTint="@color/colorPrimary"
+        android:title="btn_submit"
+        app:showAsAction="ifRoom" />
+
+
+</menu>
